Calculating median
Hi Josh - I've been playing about a bit. I've proposed an alternative method to measuring the median. I've removed the rows with NA in both the mins and hours before calculating the median (rather than turning them into 0s) as these rows are true NA responses (where no time at all was recorded), as opposed to responses with 1 for hour and NA for mins or vice versa. Not sure if that makes sense?? Wanted to test whether this would have an impact on the median, it seems as if it hasn't.